+++ packages/ordia-core/ordia/__init__.py
@@ -18,7 +18,7 @@ try:
 
     except PackageNotFoundError:
 
-        __version__ = "0.9.0"
+        __version__ = "0.9.1"
 
 except ImportError:  # pragma: no cover